backlog: move handlers to the top level
