+++ internal/server/handlers/value_handler_test.go
