+++ internal/server/handlers/values_handler_test.go
