+++ internal/server/metrics/service/service.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	"github.com/stepkareserva/obsermon/internal/models"
-	"github.com/stepkareserva/obsermon/internal/server/metrics/handlers"
+	"github.com/stepkareserva/obsermon/internal/server/handlers"
 )
 
 type Service struct {
