+++ internal/server/mocks/mock_service.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=service.go -destination=../../mocks/mock_service.go -package=mocks
+//	mockgen -source=service.go -destination=../mocks/mock_service.go -package=mocks
 //
 
 // Package mocks is a generated GoMock package.
