+++ internal/server/routing/service.go
@@ -1,6 +1,6 @@
 package routing
 
-import "github.com/stepkareserva/obsermon/internal/server/metrics/handlers"
+import "github.com/stepkareserva/obsermon/internal/server/handlers"
 
 type Service interface {
 	handlers.Service
